Revert "Add some casts for %j"
This reverts commit c4d2e483a39176a476c56d35879423fe6e33c0cd.

(Merged from https://github.com/openssl/openssl/pull/4160)

+++ apps/enc.c
@@ -555,10 +555,8 @@ int enc_main(int argc, char **argv)
 
     ret = 0;
     if (verbose) {
-        BIO_printf(bio_err, "bytes read   : %8ju\n",
-                   (uintmax_t)BIO_number_read(in));
-        BIO_printf(bio_err, "bytes written: %8ju\n",
-                   (uintmax_t)BIO_number_written(out));
+        BIO_printf(bio_err, "bytes read   : %8ju\n", BIO_number_read(in));
+        BIO_printf(bio_err, "bytes written: %8ju\n", BIO_number_written(out));
     }
  end:
     ERR_print_errors(bio_err);
